refactor: rename variables to maintain patterns
